+++ src/recoil/DeliveryInfo/atoms/selectedDeliveryInfoListState.ts
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import { DeliveryInfo } from '../../../../types/DeliveryInfo.type';
+import { DeliveryInfo } from '../../../types/DeliveryInfo.type';
 
 export const selectedDeliveryInfoListState = atom<DeliveryInfo[]>({
   key: 'selectedDeliveryInfoListState',
